Experimenting with zenburn again
I had a long period when solarized light seemed much better for reading
on the screen, but I have been experimenting a good deal of fatigue
recently. My environment at the office changed quite a bit, and it is
possible that I would benefit from darker theme now.

But I was never a fan of dark solarized, so there's that.